refactor(ChessManager.java) : ChessManager와 GameStatus 위치 수정
