+++ src/main/java/chess/controller/command/CommandMenu.java
@@ -1,6 +1,6 @@
 package chess.controller.command;
 
-import chess.manager.ChessManager;
+import chess.domain.manager.ChessManager;
 
 import java.util.Arrays;
 import java.util.List;
