+++ src/main/java/chess/controller/command/Move.java
@@ -1,7 +1,7 @@
 package chess.controller.command;
 
 import chess.domain.board.position.Position;
-import chess.manager.ChessManager;
+import chess.domain.manager.ChessManager;
 
 import java.util.List;
